Fix to keep a count for g@
Suppose that we have the following definition:

    call operator#user#define('something', '...')
    map _  <Plug>(operator-something)

When a user types

    2__

it is expanded to

    2:<C-u>call ...<Return><SID>(count)<SID>(register)g@

In old versions, <SID>(count) directly uses v:count to refer the count.
It works in old versions of Vim, but a patch between Vim 7.3.003 and Vim
7.3.285 (perhaps Vim 7.3.086) changes behavior about v:count.  v:count
seems to be reset after finishing an Ex command, so that <SID>(count)
fails to pass the actual count to g@ in newer versions of Vim.